requirements.txt
# dbt
dbt-core==1.11.2
dbt-postgres==1.8.2

# Postgres + S3 utilities
psycopg2-binary==2.9.9
boto3==1.34.162
python-dateutil==2.9.0.post0

# Quality-of-life (optional but useful)
python-dotenv>=1.0
rich>=13.7
